move gamification into its own lib
