+++ libs/server/common/src/index.ts
@@ -2,6 +2,4 @@ export * from './activity';
 export * from './core';
 export * from './event';
 export * from './core/firebase';
-export * from './gamification';
 export * from './user';
-
