+++ libs/server/gamify/README.md
@@ -1,0 +1,7 @@
+# gamify
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build gamify` to build the library.
