+++ libs/server/gamify/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "@codeheroes/gamify",
+  "version": "0.0.1",
+  "private": true,
+  "dependencies": {
+    "tslib": "^2.3.0"
+  },
+  "type": "commonjs",
+  "main": "./src/index.js",
+  "types": "./src/index.d.ts",
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    }
+  },
+  "module": "./dist/index.js"
+}
