+++ libs/server/gamify/project.json
@@ -1,0 +1,20 @@
+{
+  "name": "gamify",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/server/gamify/src",
+  "projectType": "library",
+  "tags": [],
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/server/gamify",
+        "tsConfig": "libs/server/gamify/tsconfig.lib.json",
+        "packageJson": "libs/server/gamify/package.json",
+        "main": "libs/server/gamify/src/index.ts",
+        "assets": ["libs/server/gamify/*.md"]
+      }
+    }
+  }
+}
