+++ libs/server/gamify/src/activities/base/activity-calculator.base.ts
@@ -1,4 +1,4 @@
-import { UserActivity } from '../../../activity/activity.model';
+import { UserActivity } from '@codeheroes/common';
 import { GameXpSettings, XpBreakdownItem, XpCalculationResponse } from '../../models/gamification.model';
 
 export abstract class BaseActivityCalculator {
