+++ libs/server/gamify/src/activities/pull-request/pr-processor.ts
@@ -1,10 +1,9 @@
-import { ActivityType, UserActivity } from '../../../activity/activity.model';
-import { getCurrentTimeAsISO, logger } from '../../../core/firebase';
+import { ActivityType, UserActivity, getCurrentTimeAsISO, logger } from '@codeheroes/common';
 import { calculateLevelProgress } from '../../core/level.utils';
 import { ActivityProcessingResult, XpCalculationResponse } from '../../models/gamification.model';
 import { BaseActivityProcessor } from '../base/activity-processor.base';
 
-export class IssueProcessor extends BaseActivityProcessor {
+export class PullRequestProcessor extends BaseActivityProcessor {
   async processActivity(
     userId: string,
     activityId: string,
@@ -28,8 +27,8 @@ export class IssueProcessor extends BaseActivityProcessor {
         const updatedXp = currentXp + xpResult.totalXp;
         const levelProgress = calculateLevelProgress(updatedXp, userData.achievements || [], userData.tasks || []);
 
-        // Check for issue-specific achievements
-        const processingResult = await this.processIssueAchievements(userData, activity);
+        // Check for PR-specific achievements
+        const processingResult = await this.processPullRequestAchievements(userData, activity);
 
         // Add XP calculation results
         processingResult.xp = {
@@ -45,8 +44,14 @@ export class IssueProcessor extends BaseActivityProcessor {
           currentLevelXp: levelProgress.currentLevelXp,
           xpToNextLevel: levelProgress.xpToNextLevel,
           updatedAt: getCurrentTimeAsISO(),
-          // Update issue-specific stats
-          'stats.issues': this.updateIssueStats(userData, activity),
+          //   // Update PR-specific stats
+          //   'stats.pullRequests': {
+          //     total: (userData.stats?.pullRequests?.total || 0) + 1,
+          //     merged:
+          //       activity.type === ActivityType.PR_MERGED
+          //         ? (userData.stats?.pullRequests?.merged || 0) + 1
+          //         : userData.stats?.pullRequests?.merged || 0,
+          //   },
         });
 
         // Update activity document
@@ -63,13 +68,13 @@ export class IssueProcessor extends BaseActivityProcessor {
         );
       });
 
-      logger.info('Successfully processed issue activity', {
+      logger.info('Successfully processed pull request activity', {
         userId,
         activityId,
         activityType: activity.type,
       });
     } catch (error) {
-      logger.error('Failed to process issue activity', {
+      logger.error('Failed to process pull request activity', {
         error,
         userId,
         activityId,
@@ -79,75 +84,55 @@ export class IssueProcessor extends BaseActivityProcessor {
     }
   }
 
-  private updateIssueStats(userData: any, activity: UserActivity): any {
-    const currentStats = userData.stats?.issues || {
-      total: 0,
-      closed: 0,
-      reopened: 0,
-    };
-
-    switch (activity.type) {
-      case ActivityType.ISSUE_CREATED:
-        return {
-          ...currentStats,
-          total: currentStats.total + 1,
-        };
-      case ActivityType.ISSUE_CLOSED:
-        return {
-          ...currentStats,
-          closed: currentStats.closed + 1,
-        };
-      case ActivityType.ISSUE_REOPENED:
-        return {
-          ...currentStats,
-          reopened: currentStats.reopened + 1,
-        };
-      default:
-        return currentStats;
-    }
-  }
-
-  private async processIssueAchievements(userData: any, activity: UserActivity): Promise<ActivityProcessingResult> {
+  private async processPullRequestAchievements(
+    userData: any,
+    activity: UserActivity,
+  ): Promise<ActivityProcessingResult> {
     const processingResult: ActivityProcessingResult = {
       processed: true,
       processedAt: getCurrentTimeAsISO(),
       achievements: [],
     };
 
-    const currentStats = userData.stats?.issues || { total: 0, closed: 0 };
+    // always return for now
+    if (userData) {
+      return processingResult;
+    }
 
-    // First issue created achievement
-    if (activity.type === ActivityType.ISSUE_CREATED && currentStats.total === 0) {
+    // Check for first PR achievement
+    if (!userData.stats?.pullRequests?.total) {
       processingResult.achievements?.push({
-        id: 'first_issue',
-        name: 'Issue Reporter',
-        description: 'Created your first issue',
+        id: 'first_pr',
+        name: 'First Pull Request',
+        description: 'Created your first pull request',
         progress: 100,
         completed: true,
         completedAt: getCurrentTimeAsISO(),
       });
     }
 
-    // Issue resolution achievements
-    if (activity.type === ActivityType.ISSUE_CLOSED) {
-      const closedCount = currentStats.closed + 1;
+    // Check for PR merge achievements
+    if (activity.type === ActivityType.PR_MERGED) {
+      const mergedCount = (userData.stats?.pullRequests?.merged || 0) + 1;
 
-      if (closedCount === 1) {
+      // Achievement for first merged PR
+      if (mergedCount === 1) {
         processingResult.achievements?.push({
-          id: 'first_issue_closed',
-          name: 'Problem Solver',
-          description: 'Closed your first issue',
+          id: 'first_merged_pr',
+          name: 'First Merged PR',
+          description: 'Got your first pull request merged',
           progress: 100,
           completed: true,
           completedAt: getCurrentTimeAsISO(),
         });
       }
 
-      if (closedCount === 10) {
+      // Achievement for 10 merged PRs
+      if (mergedCount === 10) {
         processingResult.achievements?.push({
-          id: 'issue_master',
-          name: 'Issue Master',
-          description: 'Closed 10 issues',
+          id: 'pr_master',
+          name: 'PR Master',
+          description: 'Got 10 pull requests merged',
           progress: 100,
           completed: true,
           completedAt: getCurrentTimeAsISO(),
