+++ tsconfig.base.json
@@ -17,6 +17,7 @@
     "paths": {
       "@codeheroes/common": ["libs/server/common/src/index.ts"],
       "@codeheroes/database-seeds": ["libs/database-seeds/src/index.ts"],
+      "@codeheroes/gamify": ["libs/server/gamify/src/index.ts"],
       "@codeheroes/migration": ["libs/migration/migration/src/index.ts"],
       "@codeheroes/migration-shared": ["libs/migration/migration-shared/src/index.ts"],
       "@shared/github-interfaces": ["libs/shared/github-interfaces/src/index.ts"]
